#164 Restructure the test folder, part2
