+++ my-webxr-app/tests/components/3Dpoint-space.test.tsx
@@ -1,10 +1,10 @@
 import ReactThreeTestRenderer, { waitFor } from '@react-three/test-renderer';
 import { XR } from '@react-three/xr';
 import { Vector3 } from 'three';
-import { PointSelectionProvider } from '../src/contexts/PointSelectionContext';
-import GraphingDataPoint from '../src/components/GraphingDataPoint';
-import createPosition from '../src/components/Positions';
-import MockServer from './MockServer';
+import { PointSelectionProvider } from '../../src/contexts/PointSelectionContext';
+import GraphingDataPoint from '../../src/components/GraphingDataPoint';
+import createPosition from '../../src/components/Positions';
+import MockServer from '../MockServer';
 
 const maxNum: number = 10;
 const scaleFactor: number = 1;
